# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AeroFS API",
    "description": "Remote file manager REST API. All endpoints require session authentication except public share endpoints.",
    "license": {
      "name": "Proprietary"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/transfers": {
      "get": {
        "tags": [
          "transfers"
        ],
        "summary": "List active and undismissed transfer jobs (scoped by user ownership and connection permissions)",
        "operationId": "list_transfers",
        "responses": {
          "200": {
            "description": "List of transfer jobs",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TransferJob"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "transfers"
        ],
        "summary": "Queue a new transfer job with full source and destination authorization",
        "operationId": "create_transfer",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransferRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Transfer job queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateTransferResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/transfers/clear-finished": {
      "post": {
        "tags": [
          "transfers"
        ],
        "summary": "Dismiss all finished transfer jobs for the authenticated user (persistent Clear)",
        "operationId": "clear_finished_transfers",
        "responses": {
          "200": {
            "description": "Finished transfers cleared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClearFinishedTransfersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/transfers/{id}/cancel": {
      "post": {
        "tags": [
          "transfers"
        ],
        "summary": "Cancel an active transfer job (enforcing user ownership)",
        "operationId": "cancel_transfer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Transfer job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transfer cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/transfers/{id}/dismiss": {
      "post": {
        "tags": [
          "transfers"
        ],
        "summary": "Dismiss a single transfer job from history (persistent)",
        "operationId": "dismiss_transfer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Transfer job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transfer dismissed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/transfers/{id}/retry": {
      "post": {
        "tags": [
          "transfers"
        ],
        "summary": "Retry or resume an interrupted or failed transfer job",
        "operationId": "retry_transfer",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Transfer job ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Transfer queued for retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransferActionResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "CookieAuth": []
          },
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClearFinishedTransfersResponse": {
        "type": "object",
        "required": [
          "success",
          "cleared",
          "message"
        ],
        "properties": {
          "cleared": {
            "type": "integer",
            "minimum": 0
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "CreateTransferRequest": {
        "type": "object",
        "required": [
          "name",
          "transfer_type",
          "source_connection_id",
          "source_path",
          "destination_connection_id",
          "destination_path"
        ],
        "properties": {
          "destination_connection_id": {
            "type": "string"
          },
          "destination_path": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "source_connection_id": {
            "type": "string"
          },
          "source_path": {
            "type": "string"
          },
          "transfer_type": {
            "$ref": "#/components/schemas/TransferType"
          }
        }
      },
      "CreateTransferResponse": {
        "type": "object",
        "required": [
          "success",
          "job_id",
          "message"
        ],
        "properties": {
          "job_id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
...
      "ErrorCategory": {
        "type": "string",
        "enum": [
          "validation",
          "authentication",
          "authorization",
          "not_found",
          "conflict",
          "transfer_cancelled",
          "payload_too_large",
          "insufficient_storage",
          "rate_limited",
          "timeout",
          "provider",
          "io",
          "security",
          "internal"
        ]
      },
...
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "category",
          "retryable",
          "message"
        ],
        "properties": {
          "category": {
            "$ref": "#/components/schemas/ErrorCategory"
          },
          "code": {
            "type": "string"
          },
          "details": {},
          "message": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "user_action": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        }
      },
...
      "TransferActionResponse": {
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          },
          "success": {
            "type": "boolean"
          }
        }
      },
      "TransferExecutionMode": {
        "type": "string",
        "description": "Execution mode for a transfer — Inline vs Background vs Resumable (§Upload-as-Transfer)",
        "enum": [
          "inline",
          "background",
          "resumable"
        ]
      },
      "TransferJob": {
        "type": "object",
        "required": [
          "id",
          "name",
          "transfer_type",
          "source_connection_id",
          "source_path",
          "destination_connection_id",
          "destination_path",
          "status",
          "phase",
          "transferred_bytes",
          "total_bytes",
          "speed_bytes_per_sec",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "checksum": {
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "destination_connection_id": {
            "type": "string"
          },
          "destination_path": {
            "type": "string"
          },
          "dismissed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "error_message": {
            "type": [
              "string",
              "null"
            ]
          },
          "eta_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "minimum": 0
          },
          "execution_mode": {
            "$ref": "#/components/schemas/TransferExecutionMode",
            "description": "Execution mode — Inline (sync HTTP), Background (queued), Resumable (checkpointed)"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "phase": {
            "$ref": "#/components/schemas/TransferPhase"
          },
          "source_connection_id": {
            "type": "string"
          },
          "source_path": {
            "type": "string"
          },
          "speed_bytes_per_sec": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "staging": {
            "$ref": "#/components/schemas/TransferStaging",
            "description": "Staging strategy — implementation detail of TransferEngine"
          },
          "status": {
            "$ref": "#/components/schemas/TransferStatus"
          },
          "total_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "transfer_type": {
            "$ref": "#/components/schemas/TransferType"
          },
          "transferred_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "user_id": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "TransferPhase": {
        "type": "string",
        "enum": [
          "preparing",
          "transferring",
          "finalizing",
          "verifying",
          "cleaning_up",
          "completed"
        ]
      },
...
      "TransferStaging": {
        "type": "string",
        "description": "Staging strategy — implementation detail of TransferEngine, not a separate subsystem",
        "enum": [
          "none",
          "local_temp",
          "provider_temp"
        ]
      },
      "TransferStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "cancellation_requested",
          "cancelled",
          "interrupted",
          "completed",
          "failed"
        ]
      },
      "TransferType": {
        "type": "string",
        "enum": [
          "copy",
          "move",
          "upload",
          "sync"
        ]
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "aerofs_session"
      }
    }
  },
...
}